fix lint properly
