+++ vm/src/utils/mod.rs
@@ -11,25 +11,6 @@ pub fn limbs_to_biguint(x: &[u32], limb_size: usize) -> BigUint {
     result
 }
 
-// little endian.
-// Warning: This function only returns the last NUM_LIMBS*LIMB_BITS bits of
-//          the input, while the input can have more than that.
-// not sure why lint complains... this is used in tests.
-#[allow(dead_code)]
-pub fn biguint_to_limbs<const NUM_LIMBS: usize>(
-    mut x: BigUint,
-    limb_size: usize,
-) -> [u32; NUM_LIMBS] {
-    let mut result = [0; NUM_LIMBS];
-    let base = BigUint::from_u32(1 << limb_size).unwrap();
-    for r in result.iter_mut() {
-        *r = (x.clone() % &base).to_u32().unwrap();
-        x /= &base;
-    }
-    assert!(x.is_zero());
-    result
-}
-
 // Use this when num_limbs is not a constant.
 // little endian.
 // Warning: This function only returns the last NUM_LIMBS*LIMB_SIZE bits of
@@ -44,3 +25,9 @@ pub fn biguint_to_limbs_vec(mut x: BigUint, limb_size: usize, num_limbs: usize) 
     assert!(x.is_zero());
     result
 }
+
+#[cfg(test)]
+mod test_utils;
+
+#[cfg(test)]
+pub use test_utils::*;
